parallax occlusion mapping
started implementing parallax occlusion mapping ... it's almost working,
but there is some bug and I think it is I am taking a bad eyePos, or
it's not in the correct space :( I am going to seek help from a real
graphic programmer.  Checking in bad code, just turn POM off and you'll
be fine for now.