Working POM!
I found this .pdf and used it to help rewrite the POM, banged on it
until I sorted out all of the issues.  I probably am not doing it as
optimally as you might in a game engine, I ended up using maya tangents,
binormals, and building the worldToTangent in the fragment shader ...
and that is something I might be doing poorly (not a trained graphics
programmer)!